Add files via upload
A buzzer is a sounding component, that is widely used in electronic devices such as calculators, electronic warning clocks, and alarms. Buzzer has two types: active and passive. An active buzzer has an oscillator inside, which will sound as long as it is supplied with power. A passive buzzer requires an external oscillator signal (generally using PWM with different frequencies) to make a sound.